+++ andcrowd/src/App.js
@@ -53,9 +53,10 @@ import AndCreateImg from './pages/and/AndCreateImg';
 import ScrollToTop from './components/ScrollToTop';
 import AndApplicantAdmin from './pages/and/AndApplicantAdmin';
 import { NaverLoginCallback } from './components/sign/NaverLoginCallback';
-import CrowdDetail from './pages/crowd/detail';
+import CrowdDetail from './pages/crowd/CrowdDetail';
 import CrowdList from './pages/CrowdList';
 import CrowdCreate from './pages/crowd/CrowdCreate';
+import LoginRoute from './components/route/LoginRoute';
 
 const sections = [
   { title: '홈', url: '/' },
@@ -90,79 +91,81 @@ function App() {
         <div className="main-content">
           <ScrollToTop></ScrollToTop>
           <Routes>
-            {/* 로그인되지 않은 상태에서만 접근 가능 */}
-            <Route element={<PublicRoute />}>
-              <Route path="/login" element={<Login />} />
-              <Route path="/signup" element={<Signup />} />
-              <Route path="/naverLoginCallback/*" element={<NaverLoginCallback />} />
-            </Route>
+            <Route element={<LoginRoute />}>
+              {/* 로그인되지 않은 상태에서만 접근 가능 */}
+              <Route element={<PublicRoute />}>
+                <Route path="/login" element={<Login />} />
+                <Route path="/signup" element={<Signup />} />
+                <Route path="/naverLoginCallback/*" element={<NaverLoginCallback />} />
+              </Route>
 
-            {/* 누구라 접근 가능 */}
-            {/* User 관련 */}
-            <Route path="/test" element={<Test />} />
-            <Route path="/user/2/and" element={<Test />} />
-            <Route path="/" element={<Home />} />
-            <Route path="/user/:userId" element={<MyPage />} />
+              {/* 누구라 접근 가능 */}
+              {/* User 관련 */}
+              <Route path="/test" element={<Test />} />
+              <Route path="/user/2/and" element={<Test />} />
+              <Route path="/" element={<Home />} />
+              <Route path="/user/:userId" element={<MyPage />} />
 
-            {/* And 관련 */}
-            <Route path="/and/list" element={<AndList />} />
-            <Route path="/and/scroll" element={<AndScroll />} />
-            <Route path="/and/:andId" element={<AndDetail />} />
-            <Route path="/and/:andId/qna/list" element={<AndQna />} />
-            <Route path="/and/:andId/qna/:andQnaId" element={<AndQnaDetail />} />
-            <Route path="/and/:andId/qna/:andQnaId/update" element={<AndQnaUpdate />} />
-            <Route path="/and/:andId/qna/reply/:andQnaId/:andQnaReplyId/update" element={<AndReplyUpdate />} />
-            <Route path="/and/:andId/applicant/list" element={<AndApplicant />} />
-            <Route path="/and/:andId/applicant/:andApplyId/update" element={<AndApplicantUpdate />} />
-            <Route path="/and/:andId/applicant/:andApplyId" element={<AndApplicantDetail />} />
-            <Route path="/and/:andId/applicant/:andApplyId/admin" element={<AndApplicantAdmin />} />
-            <Route path="/and/:andId/board/list" element={<AndBoard />} />
-            <Route path="/and/:andId/board/:andBoardId" element={<AndBoardDetail />} />
-            <Route path="/and/:andId/role/list" element={<AndRole />} />
-            <Route path="/and/:andId/role/:andRoleId" element={<AndRoleDetail />} />
+              {/* And 관련 */}
+              <Route path="/and/list" element={<AndList />} />
+              <Route path="/and/scroll" element={<AndScroll />} />
+              <Route path="/and/:andId" element={<AndDetail />} />
+              <Route path="/and/:andId/qna/list" element={<AndQna />} />
+              <Route path="/and/:andId/qna/:andQnaId" element={<AndQnaDetail />} />
+              <Route path="/and/:andId/qna/:andQnaId/update" element={<AndQnaUpdate />} />
+              <Route path="/and/:andId/qna/reply/:andQnaId/:andQnaReplyId/update" element={<AndReplyUpdate />} />
+              <Route path="/and/:andId/applicant/list" element={<AndApplicant />} />
+              <Route path="/and/:andId/applicant/:andApplyId/update" element={<AndApplicantUpdate />} />
+              <Route path="/and/:andId/applicant/:andApplyId" element={<AndApplicantDetail />} />
+              <Route path="/and/:andId/applicant/:andApplyId/admin" element={<AndApplicantAdmin />} />
+              <Route path="/and/:andId/board/list" element={<AndBoard />} />
+              <Route path="/and/:andId/board/:andBoardId" element={<AndBoardDetail />} />
+              <Route path="/and/:andId/role/list" element={<AndRole />} />
+              <Route path="/and/:andId/role/:andRoleId" element={<AndRoleDetail />} />
 
-            {/* 로그인된 유저만 접근 가능 */}
-            <Route element={<PrivateRoute />}>
-            <Route path="/and/create" element={<AndCreate />} />
-
-            <Route path="/and/:andId/img/create" element={<AndCreateImg />} />
-
-              <Route path="/and/:andId/update" element={<AndUpdate />} />
-              <Route path="/and/:andId/qna/create" element={<AndQnaCreate />} />
+              {/* 로그인된 유저만 접근 가능 */}
+              <Route element={<PrivateRoute />}>
               <Route path="/and/create" element={<AndCreate />} />
-              <Route path="/and/:andId/qna/reply/:andQnaId/create" element={<AndReplyCreate />} />
-              <Route path="/and/:andId/applicant/create" element={<AndApplicantCreate />} />
-              <Route path="/and/:andId/board/create" element={<AndBoardCreate />} />
-              <Route path="/and/:andId/board/:andBoardId/update" element={<AndBoardUpdate />} />
-              <Route path="/and/:andId/member/create" element={<AndMemberCreate />} />
-              <Route path="/and/:andId/member/:memberId/update" element={<AndMemberUpdate />} />
-              <Route path="/and/:andId/role/create" element={<AndRoleCreate />} />
-              <Route path="/and/:andId/role/:andRoleId/update" element={<AndRoleUpdate />} />
+
+              <Route path="/and/:andId/img/create" element={<AndCreateImg />} />
+
+                <Route path="/and/:andId/update" element={<AndUpdate />} />
+                <Route path="/and/:andId/qna/create" element={<AndQnaCreate />} />
+                <Route path="/and/create" element={<AndCreate />} />
+                <Route path="/and/:andId/qna/reply/:andQnaId/create" element={<AndReplyCreate />} />
+                <Route path="/and/:andId/applicant/create" element={<AndApplicantCreate />} />
+                <Route path="/and/:andId/board/create" element={<AndBoardCreate />} />
+                <Route path="/and/:andId/board/:andBoardId/update" element={<AndBoardUpdate />} />
+                <Route path="/and/:andId/member/create" element={<AndMemberCreate />} />
+                <Route path="/and/:andId/member/:memberId/update" element={<AndMemberUpdate />} />
+                <Route path="/and/:andId/role/create" element={<AndRoleCreate />} />
+                <Route path="/and/:andId/role/:andRoleId/update" element={<AndRoleUpdate />} />
+                
+                <Route path="/crowd/create" element={<CrowdCreate />} />
+                <Route path="/crowd/:crowdId/board/:crowdBoardId/update" element={<CrowdBoardUpdate />} />
+              </Route>
               
-              <Route path="/crowd/create" element={<CrowdCreate />} />
-              <Route path="/crowd/:crowdId/board/:crowdBoardId/update" element={<CrowdBoardUpdate />} />
+              <Route path="/and/:andId/member/list" element={<AndMember />} />
+              <Route path="/and/:andId/member/:memberId" element={<AndMemberDetail />} />
+
+              <Route path="/and/:andId/chat" element={<AndChat />} />
+              <Route path="/and/:andId/chat/room/:roomId/name-update" element={<ChatroomUpdate />} />
+
+              <Route path="/crowd/:crowdId/board/all" element={<CrowdBoardList />} />
+              <Route path="/crowd/:crowdId/board/:crowdBoardId" element={<CrowdBoardDetail />} />
+              <Route path="/crowd/:crowdId/insert" element={<CrowdBoardInsert />} />
+              {/* <Route path="/crowd/:crowdId/reward/:rewardId/payment" element={<CrowdPayment />} /> */}
+              <Route path="/crowd/detail/:crowdId" element={<CrowdDetail />} />
+              <Route path="/crowd/:crowdId/reward/:rewardId/payment" element={<CrowdPaymentFunction />} />
+              <Route path=":projectType/:projectId/ad/payment" element={<AdPaymentFunction />} />
+              <Route path="/signup" element={<Signup />} />
+              <Route path="/crowd/list" element={<Login />} />
+              <Route path="/crowd/create" element={<Login />} />
+              <Route path="/team" element={<Login />} />
+              <Route path="/help" element={<Login />} />
+              <Route path="*" element={<NotFound />} />
+              <Route path='/chat' element={<ChatPage />} />
             </Route>
-            
-            <Route path="/and/:andId/member/list" element={<AndMember />} />
-            <Route path="/and/:andId/member/:memberId" element={<AndMemberDetail />} />
-
-            <Route path="/and/:andId/chat" element={<AndChat />} />
-            <Route path="/and/:andId/chat/room/:roomId/name-update" element={<ChatroomUpdate />} />
-
-            <Route path="/crowd/:crowdId/board/all" element={<CrowdBoardList />} />
-            <Route path="/crowd/:crowdId/board/:crowdBoardId" element={<CrowdBoardDetail />} />
-            <Route path="/crowd/:crowdId/insert" element={<CrowdBoardInsert />} />
-            <Route path="/crowd/:crowdId/reward/:rewardId/payment" element={<CrowdPayment />} />
-            <Route path="/crowd/detail/:crowdId" element={<CrowdDetail />} />
-            <Route path="/crowd/:crowdId/reward/:rewardId/payment" element={<CrowdPaymentFunction />} />
-            <Route path=":projectType/:projectId/ad/payment" element={<AdPaymentFunction />} />
-            <Route path="/signup" element={<Signup />} />
-            <Route path="/crowd/list" element={<Login />} />
-            <Route path="/crowd/create" element={<Login />} />
-            <Route path="/team" element={<Login />} />
-            <Route path="/help" element={<Login />} />
-            <Route path="*" element={<NotFound />} />
-            <Route path='/chat' element={<ChatPage />} />
           </Routes>
         </div>
       </div>
