+++ andcrowd/src/components/user/GetUserInfo.js
@@ -1,14 +1,19 @@
 export const GetUserInfo = (userId, setUserInfo) => {
-    fetch(`/user/${userId}`, {
-        method: "GET",
-        headers:{
-          'Authorization': `Bearer ${localStorage.getItem("access_token")}`,
-          'Content-': 'application/json; text=utf-8'
-        }
-    }).then(res => {
-        if(!res.ok) throw new Error(`/user/${userId}: ${res}`);
-        return res.json();
-    }).then(data => {
-        setUserInfo(data);
-    });
+    try{
+        fetch(`/user/${userId}`, {
+            method: "GET",
+            headers:{
+              'Authorization': `Bearer ${localStorage.getItem("access_token")}`,
+              'Content-Type': 'application/json; text=utf-8'
+            }
+        }).then(res => {
+            if(!res.ok) throw new Error(`/user/${userId}: ${res}`);
+            return res.json();
+        }).then(data => {
+            setUserInfo(data);
+        });
+    } catch(error){
+        console.error("GetUserInfo: " + error);
+    }
+    
 }
